// src/models/UserWithTokensModel.ts
export interface IUserWithTokensModel{
    id: number;
    username: string;
    email: string
    firstName: string;
    lastName: string;
    gender: string;
    image: string;
    accessToken: string; // JWT accessToken (for backward compatibility) in response and cookies
    refreshToken: string
}